Login Almost Completed and other modifications were done
i did a lot of changes since i was getting a lot of errors setting SQL and loading FXML files and the only solution was to put every class in a package and to put FXML in the "resources" folder. sorry for the inconvenience and i will setup SQL for you later so you have the full functional program.