Finishing the reservelabsclasses for nearly
There are some problems in this one, because it is a choice box to choose rather you want to reserve a lab ore class, there are to you can't add more than 1 class and 1 lab for some reason. I tried to change the comprasion to be between lab numbers, but the error still comes as dublicated data type.